// self-triage-app/src/types/HelperEnums/EResult.ts
export enum EResult {
    FamilyDoctor = "palun kontakteeruge perearstiga.",
    EmergencyRoom = "palun pöörduge EMO-sse või helistage numbrile 112.",
    Pharmacy = "palun pöörduke apteeki.",
    OnlyEmergencyRoom = "palun pöörduge EMO-sse.",
    ToothPieceLoss = "Otsi tükk üles. Võta hambatükk kaasa ja mine kiiresti hambaarsti juurde",
    ToothCompleteLoss = "Otsi hammas üles. Ära katsu hambajuure osa. Kui leitud hammas on määrdunud, siis loputa korraks puhta veega. Edasi tegutse vastavalt kuidas sulle sobib. Pane hammas oma kohale tagasi. pane hammas piima või 0,9NaCl lahusesse (füsioloogiline lahus) hoida hammas oma põses või topsis, sülje sees. Pöördu hambarsti juurde vähemalt kahe tunni jooksul"
}